refactor: extract runtime controller
